+++ setup.cfg
@@ -14,7 +14,7 @@ max-complexity = 10
 max-line-length = 150
 
 [tool:pytest]
-testpaths = src/tests
+testpaths = tests
 python_files = test_*.py
 asyncio_mode = auto
 filterwarnings =
