+++ tests/10_utils/test_decode_alliance_membership.py
