+++ tests/10_utils/test_encode_alliance_membership.py
