+++ tests/40_database/test_crud_delete_collection.py
