+++ tests/40_database/test_crud_get_alliance_from_collection.py
