+++ tests/40_database/test_crud_has_alliance_history.py
