+++ tests/40_database/test_crud_has_user_history.py
