+++ tests/50_endpoints/test_api_delete_collection.py
