+++ tests/50_endpoints/test_api_get_alliance_from_collection.py
