+++ tests/50_endpoints/test_api_get_alliances_from_collection.py
