+++ tests/50_endpoints/test_api_get_collections.py
